+++ src/utils.rs
@@ -50,8 +50,7 @@ pub(crate) fn buffer_to_string<F>(context: F, buffer: Vec<u8>) -> Result<String,
 where
     F: FnOnce() -> String,
 {
-    String::from_utf8(buffer)
-        .map_err(|error| Error::utf8_conversion_error(error, context()))
+    String::from_utf8(buffer).map_err(|error| Error::utf8_conversion_error(error, context()))
 }
 
 #[cfg(test)]