+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow rendering to `std::io::Write`
 - Follow symlinks in glob
 - Allow including lists of templates
+- Comment tags can now use whitespace control
 
 ## 1.6.1 (2020-12-29)